Separated DrawerMenu into a different file.
