+++ src/components/Nav/OpenSimAppBar.tsx
@@ -3,7 +3,6 @@ import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import Toolbar from '@mui/material/Toolbar';
 import IconButton from '@mui/material/IconButton';
-import Typography from '@mui/material/Typography';
 import Link from '@mui/material/Link';
 import logo_dark from './logo-dark.svg';
 import logo from './logo.svg';
